Fix streams to handle huge data allocations

+++ src/workgroups.c
@@ -225,8 +225,8 @@ int manage_streams(RuntimeWorkgroupConfig* wg, RuntimeConfig* runcfg)
                 bstring t = bstrcpy(istream->dims->entry[k]);
                 DEBUG_PRINT(DEBUGLEV_DEVELOP, Stream %d: dimsize before %s, j, bdata(t));
                 replace_all(runcfg->global_results, t, NULL);
-                int res = 0;
-                int c = sscanf(bdata(t), "%d", &res);
+                size_t res = 0;
+                int c = sscanf(bdata(t), "%ld", &res);
                 if (c == 1)
                 {
                     ostream->dimsizes[k] = res;
